Fixed an issue with ship rotation while moving.
Turns out since my cursor was not in worldspace and the camera movement slightly behind/ahead of the rest of the game, it causes tiny miscalulations that lead to huge angle errors.

It should all be working in world space now.

Do I update the camera "setScroll()" code to keep things consistent?